+++ packages/react/src/evm/index.ts
@@ -33,7 +33,10 @@ export { useMetamask } from "./hooks/wallets/useMetamask";
 export { useCoinbaseWallet } from "./hooks/wallets/useCoinbaseWallet";
 export { useFrameWallet } from "./hooks/wallets/useFrame";
 export { useBloctoWallet } from "./hooks/wallets/useBloctoWallet";
-export { useEmbeddedWallet } from "./hooks/wallets/useEmbeddedWallet";
+export {
+  useEmbeddedWallet,
+  useEmbeddedWalletUserEmail,
+} from "./hooks/wallets/useEmbeddedWallet";
 export { useEmbeddedWalletSendVerificationEmail } from "./hooks/useEmbeddedWalletSendVerificationEmail";
 
 export {
@@ -71,4 +74,5 @@ export {
   setWalletAnalyticsEnabled,
   CoreWallet,
   DefiWallet,
+  Coin98Wallet,
 } from "@thirdweb-dev/wallets";