+++ packages/react/src/evm/utils/ifWaiting.ts
@@ -2,7 +2,7 @@
  * if the `options.for` promise takes longer than the `options.moreThan` time, `options.do` callback function will be called
  *
  * this is useful to prevent showing the loading animation for a super short time that looks like flickering
- * @param options
+ * @param options - Options for the ifWaiting function
  */
 export async function ifWaiting(options: {
   for: Promise<any>;